+++ src/main/java/com/sparta/ezpzuser/domain/cart/dto/CartUpdateRequestDto.java
@@ -9,5 +9,5 @@ import lombok.NoArgsConstructor;
 public class CartUpdateRequestDto {
 
     @NotNull(message = "장바구니에 담을 상품의 수량을 입력해주세요.")
-    Long quantity;
+    int quantity;
 }
